src/requirements.txt
neat-python==0.92
numpy==1.17.2
pip==10.0.1
pygame==1.9.6
setuptools==39.1.0